fix(baselines): pass explicit cache_position to model.generate
transformers 4.50+ rewrote prepare_inputs_for_generation so that when a
pre-populated past_key_values is provided without cache_position, it
derives an empty cache_position tensor and crashes on `cache_position[-1]`.
Pass the correct cache_position (positions [chunk_len .. chunk_len+new_len))
explicitly from both full_reuse_generate and cacheblend_generate.

Also revert dtype= -> dtype= (4.57 supports it) and pin the Modal image
to transformers==4.57.0 since DynamicCache.layers[*].keys API requires 4.55+.

+++ workspace/code/eval/run_eval.py
@@ -62,7 +62,7 @@ def _load_model(cfg: Dict[str, Any]):
     dtype = getattr(torch, cfg["model"].get("dtype", "float16"))
     device = cfg["model"].get("device", "cuda" if torch.cuda.is_available() else "cpu")
     tokenizer = AutoTokenizer.from_pretrained(name)
-    model = AutoModelForCausalLM.from_pretrained(name, torch_dtype=dtype).to(device)
+    model = AutoModelForCausalLM.from_pretrained(name, dtype=dtype).to(device)
     model.eval()
     return model, tokenizer, device, dtype
